+++ src/redturtle/prenotazioni/behaviors/booking_folder.py
@@ -4,7 +4,8 @@ from plone.dexterity.interfaces import IDexterityContent
 from plone.supermodel import model
 from zope import schema
 from zope.component import adapter
-from zope.interface import implementer, provider
+from zope.interface import implementer
+from zope.interface import provider
 
 from redturtle.prenotazioni import _