+++ src/redturtle/prenotazioni/browser/prenotazioni_context_state.py
@@ -2,23 +2,31 @@
 import itertools
 import json
 import logging
-from datetime import date, datetime, timedelta
+from datetime import date
+from datetime import datetime
+from datetime import timedelta
 
 import six
 from DateTime import DateTime
 from plone import api
 from plone.memoize.view import memoize
 from Products.Five.browser import BrowserView
-from six.moves import map, range
+from six.moves import map
+from six.moves import range
 
-from redturtle.prenotazioni import _, get_or_create_obj, tznow
+from redturtle.prenotazioni import _
+from redturtle.prenotazioni import get_or_create_obj
+from redturtle.prenotazioni import tznow
 from redturtle.prenotazioni.adapters.booker import IBooker
 from redturtle.prenotazioni.adapters.conflict import IConflictManager
-from redturtle.prenotazioni.adapters.slot import BaseSlot, ISlot
-from redturtle.prenotazioni.config import PAUSE_PORTAL_TYPE, PAUSE_SLOT
+from redturtle.prenotazioni.adapters.slot import BaseSlot
+from redturtle.prenotazioni.adapters.slot import ISlot
+from redturtle.prenotazioni.config import PAUSE_PORTAL_TYPE
+from redturtle.prenotazioni.config import PAUSE_SLOT
 from redturtle.prenotazioni.content.pause import Pause
 from redturtle.prenotazioni.content.prenotazione_type import PrenotazioneType
-from redturtle.prenotazioni.utilities.dateutils import hm2DT, hm2seconds
+from redturtle.prenotazioni.utilities.dateutils import hm2DT
+from redturtle.prenotazioni.utilities.dateutils import hm2seconds
 from redturtle.prenotazioni.utilities.urls import urlify
 
 logger = logging.getLogger(__name__)
@@ -1020,23 +1028,6 @@ class PrenotazioniContextState(BrowserView):
             return
         good_slots.sort(key=lambda x: x.lower_value)
         return good_slots[0]
-
-    def get_less_used_gates(self, booking_date, available_gates):
-        """
-        Find which gate is les busy the day of the booking
-        """
-        availability = self.get_free_slots(booking_date)
-        # Create a dictionary where keys is the time the gate is free, and
-        # value is a list of gates
-        free_time_map = {}
-        for gate, free_slots in six.iteritems(availability):
-            if gate not in available_gates:
-                continue
-            free_time = sum(map(BaseSlot.__len__, free_slots))
-            free_time_map.setdefault(free_time, []).append(gate)
-        # Get a random choice among the less busy one
-        max_free_time = max(free_time_map.keys())
-        return free_time_map[max_free_time]
 
     def __call__(self):
         """Return itself"""