+++ src/redturtle/prenotazioni/events/prenotazione.py
@@ -1,17 +1,16 @@
 # -*- coding: utf-8 -*-
 import hashlib
-from email.utils import formataddr, parseaddr
+from email.utils import formataddr
+from email.utils import parseaddr
 from logging import getLogger
 
 from plone import api
-from plone.app.event.base import default_timezone
 from plone.registry.interfaces import IRegistry
-from plone.stringinterp.interfaces import IStringSubstitution
 from Products.CMFPlone.interfaces.controlpanel import IMailSchema
-from zope.component import getAdapter, getMultiAdapter, getUtility
-from zope.i18n import translate
+from zope.component import getMultiAdapter
+from zope.component import getUtility
 
-from redturtle.prenotazioni import _, is_migration
+from redturtle.prenotazioni import is_migration
 from redturtle.prenotazioni.adapters.booker import IBooker
 from redturtle.prenotazioni.interfaces import IPrenotazioneEmailMessage
 
@@ -50,7 +49,6 @@ def notify_on_after_transition_event(context, event):
     """The messages are being send only if the following flags on the PrenotazioniFolder are set"""
     if is_migration():
         return
-
     booking_folder = context.getPrenotazioniFolder()
     flags = {
         i: getattr(booking_folder, f"notify_on_{i}", False)
@@ -129,62 +127,16 @@ def send_email_to_managers(booking, event):
 
     if booking.isVacation():
         return
+    if not getattr(booking, "email_responsabile", []):
+        return
 
-    booking_folder = booking.getPrenotazioniFolder()
+    adapter = getMultiAdapter(
+        (booking, event), IPrenotazioneEmailMessage, name="notify_manager"
+    )
 
-    booking_operator_url = getAdapter(
-        booking, IStringSubstitution, "booking_operator_url"
-    )()
-
-    email_list = getattr(booking_folder, "email_responsabile", "")
-    if email_list:
-        mail_template = api.content.get_view(
-            name="manager_notification_mail",
-            context=booking,
-            request=booking.REQUEST,
-        )
-        booking_date = getattr(booking, "booking_date", None)
-        parameters = {
-            "company": getattr(booking, "company", ""),
-            "booking_folder": booking_folder.title,
-            "booking_url": booking_operator_url,
-            "booking_date": booking_date.astimezone(
-                default_timezone(as_tzinfo=True)
-            ).strftime("%d/%m/%Y"),
-            "booking_hour": booking_date.astimezone(
-                default_timezone(as_tzinfo=True)
-            ).strftime("%H:%M"),
-            "booking_expiration_date": getattr(booking, "booking_expiration_date", ""),
-            "description": getattr(booking, "description", ""),
-            "email": getattr(booking, "email", ""),
-            "fiscalcode": getattr(booking, "fiscalcode", ""),
-            "gate": getattr(booking, "gate", ""),
-            "phone": getattr(booking, "phone", ""),
-            "staff_notes": getattr(booking, "staff_notes", ""),
-            "booking_type": getattr(booking, "booking_type", ""),
-            "title": getattr(booking, "title", ""),
-        }
-        mail_text = mail_template(**parameters)
-
-        mailHost = api.portal.get_tool(name="MailHost")
-        subject = translate(
-            _(
-                "new_booking_admin_notify_subject",
-                default="New booking for ${context}",
-                mapping={"context": booking_folder.title},
-            ),
-            context=booking.REQUEST,
-        )
-        for mail in email_list:
-            if mail:
-                mailHost.send(
-                    mail_text,
-                    mto=mail,
-                    mfrom=get_mail_from_address(),
-                    subject=subject,
-                    charset="utf-8",
-                    msg_type="text/html",
-                )
+    if adapter:
+        if adapter.message:
+            send_email(adapter.message)
 
 
 def set_booking_code(booking, event):