+++ src/redturtle/prenotazioni/prenotazione_event.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from zope.interface import implementer
-from zope.interface.interfaces import IObjectEvent, ObjectEvent
+from zope.interface.interfaces import IObjectEvent
+from zope.interface.interfaces import ObjectEvent
 
 
 class IMovedPrenotazione(IObjectEvent):