+++ src/redturtle/prenotazioni/upgrades.py
@@ -11,7 +11,8 @@ from plone.app.event.base import default_timezone
 from plone.app.upgrade.utils import loadMigrationProfile
 from plone.app.workflow.remap import remap_workflow
 from plone.contentrules.engine.interfaces import IRuleStorage
-from zope.component import getUtility, queryUtility
+from zope.component import getUtility
+from zope.component import queryUtility
 
 from redturtle.prenotazioni import _
 from redturtle.prenotazioni.events.prenotazione import set_booking_code
@@ -303,10 +304,8 @@ def to_1600_popolate_templates(context):
 
 
 def to_1600_upgrade_contentrules(context):
-    from plone.contentrules.engine.interfaces import (
-        IRuleAssignmentManager,
-        IRuleStorage,
-    )
+    from plone.contentrules.engine.interfaces import IRuleAssignmentManager
+    from plone.contentrules.engine.interfaces import IRuleStorage
 
     rules_to_delete = [
         "booking-accepted",
@@ -482,7 +481,9 @@ def to_2000(context):
                 title=booking_type.get("name"),
                 duration=booking_type.get("duration"),
                 container=obj,
+                requirements=getattr(obj, "cosa_serve", None),
             )
             if not booking_type.get("hidden", False):
                 api.content.transition(obj=booking_type_obj, transition="publish")
             booking_type_obj.reindexObject(idxs=["review_state"])
+        delattr(obj, "cosa_serve")