+++ src/redturtle/prenotazioni/vocabularies/requirable_booking_fields.py
@@ -2,7 +2,8 @@
 from zope.i18n import translate
 from zope.interface import implementer
 from zope.schema.interfaces import IVocabularyFactory
-from zope.schema.vocabulary import SimpleTerm, SimpleVocabulary
+from zope.schema.vocabulary import SimpleTerm
+from zope.schema.vocabulary import SimpleVocabulary
 
 from redturtle.prenotazioni import _
 from redturtle.prenotazioni.config import REQUIRABLE_AND_VISIBLE_FIELDS